refactor: lombok annotation 적용
